Add Linux-like modular build system with vmzirvium
Major Features:
- vmzirvium compressed kernel image (like Linux's vmlinuz)
- Kconfig-based configuration system
- Interactive menuconfig (scripts/menuconfig.sh)
- Modular compilation based on .config
- Gzip compression (45% size reduction: 6.2KB → 3.4KB)
- Linux-like Makefile with proper targets

Build Products:
- vmzirvium: Compressed kernel image (3.4 KB)
- zirvium: Uncompressed ELF kernel (13 KB)
- zirvium.iso: Bootable ISO with GRUB

Configuration System:
- .config: Kernel configuration file
- Kconfig: Configuration option definitions
- make menuconfig: Interactive TUI for configuration

New Make Targets:
- make vmzirvium: Build compressed kernel
- make menuconfig: Configure kernel options
- make iso: Build bootable ISO
- make run: Test in QEMU
- make help: Show all available targets
- make mrproper: Clean everything including config

Modular Structure:
- Conditional compilation based on CONFIG_* options
- Separate modules for arch, mm, proc, fs, syscall, drivers
- Each module can be enabled/disabled via menuconfig

Documentation:
- docs/BUILD_SYSTEM.md: Complete build system guide

This brings Zirvium's build system in line with Linux kernel
conventions, making it familiar to kernel developers.

+++ kernel/console.c
@@ -5,7 +5,7 @@
 #define VGA_HEIGHT 25
 #define VGA_MEMORY 0xB8000
 
-static uint16_t *vga_buffer = (uint16_t*)VGA_MEMORY;
+static uint16_t *vga_buffer = (uint16_t*)((uintptr_t)VGA_MEMORY);
 static int cursor_x = 0;
 static int cursor_y = 0;
 static uint8_t color = 0x07;
